Cofigurando DTO, Aplicando exceção para buscarId e aplicando anotação @Valid

+++ recirende/src/main/java/com/miniprojeto/trocadepontos/dto/UsuarioResponse.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class UsuarioResponse {
 
     private String nome;
-    private Integer pontuacao;
+    private BigDecimal pontuacao;
 
 }
